BIM: move Arch wall tests to their own module
