+++ src/Mod/BIM/CMakeLists.txt
@@ -27,6 +27,7 @@ SET(Arch_SRCS
     ArchSpace.py
     ArchRebar.py
     TestArch.py
+    TestArchWall.py
     ArchFrame.py
     ArchPanel.py
     ArchEquipment.py
