+++ src/Mod/BIM/TestArch.py
@@ -393,51 +393,6 @@ class ArchTest(unittest.TestCase):
             App.newDocument("ArchTest")
         App.setActiveDocument("ArchTest")
 
-    def testWall(self):
-        App.Console.PrintLog ('Checking Arch Wall...\n')
-        l=Draft.makeLine(App.Vector(0,0,0),App.Vector(-2,0,0))
-        w = Arch.makeWall(l)
-        self.assertTrue(w,"Arch Wall failed")
-
-    def testWallMultiMatAlign(self):
-        App.Console.PrintLog ('Checking Arch Wall with MultiMaterial and 3 alignments...\n')
-        matA = Arch.makeMaterial()
-        matB = Arch.makeMaterial()
-        matMulti = Arch.makeMultiMaterial()
-        matMulti.Materials = [matA, matB]
-        matMulti.Thicknesses = [100, 200] # total width different from default 200
-        pts = [App.Vector(   0,    0, 0),
-               App.Vector(1000,    0, 0),
-               App.Vector(1000, 1000, 0),
-               App.Vector(2000, 1000, 0)]
-        # wall based on wire:
-        wire = Draft.makeWire(pts)
-        wallWire = Arch.makeWall(wire)
-        wallWire.Material = matMulti
-        # wall based on sketch:
-        sketch = App.activeDocument().addObject('Sketcher::SketchObject','Sketch')
-        sketch.addGeometry([Part.LineSegment(pts[0], pts[1]),
-                            Part.LineSegment(pts[1], pts[2]),
-                            Part.LineSegment(pts[2], pts[3])])
-        wallSketch = Arch.makeWall(sketch)
-        wallSketch.Material = matMulti
-
-        alignLst = ["Left", "Center", "Right"]
-        checkLst = [[App.Vector(0, -300, 0), App.Vector(2000, 1000, 0)],
-                    [App.Vector(0, -150, 0), App.Vector(2000, 1150, 0)],
-                    [App.Vector(0,    0, 0), App.Vector(2000, 1300, 0)]]
-        for i in range(3):
-            wallWire.Align = alignLst[i]
-            wallSketch.Align = alignLst[i]
-            App.ActiveDocument.recompute()
-            for box in [wallWire.Shape.BoundBox, wallSketch.Shape.BoundBox]:
-                ptMin = App.Vector(box.XMin, box.YMin, 0)
-                self.assertTrue(ptMin.isEqual(checkLst[i][0], 1e-8),
-                                "Arch Wall with MultiMaterial and 3 alignments failed")
-                ptMax = App.Vector(box.XMax, box.YMax, 0)
-                self.assertTrue(ptMax.isEqual(checkLst[i][1], 1e-8),
-                                "Arch Wall with MultiMaterial and 3 alignments failed")
-
     def testStructure(self):
         App.Console.PrintLog ('Checking BIM Structure...\n')
         s = Arch.makeStructure(length=2,width=3,height=5)
